requirements.txt
twilio==9.0.0
openai==1.3.0
flask==2.3.3
python-dotenv==1.0.0
python-socketio==5.8.0
eventlet==0.33.3
flask-socketio==5.3.6
gunicorn==21.2.0
